collision debugger

+++ src/engine/input/Input.cpp
@@ -5,7 +5,7 @@ Input::Input()
 
 }
 
-static glm::ivec2 mousePos;
+static glm::vec2 mousePos;
 static std::map<int, bool> keys;
 static std::map<int, bool> mouse;
 
@@ -33,10 +33,10 @@ bool Input::isKeyDown(int key) {
     return keys[key];
 }
 
-glm::ivec2 Input::getMousePosition() {
+glm::vec2 Input::getMousePosition() {
     return mousePos;
 }
 
-void Input::setMousePosition(glm::ivec2 mp) {
+void Input::setMousePosition(glm::vec2 mp) {
     mousePos = mp;
 }
